Prevent increasing user set speeds
Due to the algorithm excluding certain (ideal for the calculations) speeds, we may run into the scenario where the new speed is higher than the previous one. For this version, we leave the old speed as is. Increasing user set speeds can lead from unpleasantries all the way to physical damage.